fix: el usuario solo puede editar su perfil

+++ src/main/java/com/example/adoptr_backend/service/impl/ProfileServiceImpl.java
@@ -67,21 +67,28 @@ public class ProfileServiceImpl implements ProfileService {
 
 
     @Override
-    public ProfileDTO update(Long id, ProfileDTOin dto){
+    public ProfileDTO update(Long id, ProfileDTOin dto) {
+        Long userId = AuthSupport.getUserId();
         Profile profile = getProfile(id);
+        if (!profile.getUser().getId().equals(userId)) {
+            throw new BadRequestException(Error.PROFILE_UPDATE_NOT_FOUND);
+        }
         Profile profileUpdated = ProfileMapper.MAPPER.toEntity(dto);
         ProfileMapper.MAPPER.update(profile, profileUpdated);
         Locality locality = getLocality(dto);
         profile.setLocality(locality);
         profile = profileRepository.save(profile);
+
         if (dto.getImage() != null) {
             imageService.deleteImage(profile.getId(), ImageType.PROFILE);
             Long imageId = imageService.uploadImage(dto.getImage(), ImageType.PROFILE, profile.getId());
             profile.setImageId(imageId);
             profileRepository.save(profile);
         }
+
         return ProfileMapper.MAPPER.toDto(profile);
     }
+
 
     private Profile getProfile(Long id) {
         Optional<Profile> profileOptional = profileRepository.findById(id);
